# Asymptotic speedup of ud2gf