--- a/UD2GF.hs
+++ b/UD2GF.hs
@@ -16,7 +16,8 @@ import Data.Char
 import Data.Maybe
 import Text.PrettyPrint (render)
 
-import Debug.Trace (trace)
+import Debug.Trace (trace, traceShowId, traceShow)
+import Data.Function (on)
 
 ---------
 -- to debug
@@ -40,13 +41,13 @@ test opts env string = do
   return ()
 
 showUD2GF opts env sentence = do
-  
+
   ifOpt opts "ud" $ prt sentence
 
   case errors sentence of
     [] -> return ()
     errs -> ifOpt opts "err" $ unlines errs
- 
+
   let udtree = udSentence2tree sentence
   ifOpt opts "ut" $ prUDTree udtree
 
@@ -55,7 +56,7 @@ showUD2GF opts env sentence = do
 
   let devtree1 = analyseWords env devtree0
   ifOpt opts "dt1" $ prLinesRTree (prDevNode 2) devtree1
-  
+
   let devtree = combineTrees env devtree1
   ifOpt opts "dt" $ prLinesRTree (prDevNode 4) devtree
 
@@ -64,7 +65,7 @@ showUD2GF opts env sentence = do
 
   let besttree = addBackups besttree0
   ifOpt opts "bt" $ prLinesRTree (prDevNode 1) besttree
-  
+
   let ts0 = devtree2abstrees besttree
   ifOpt opts "at0" $ unlines $ map prAbsTree ts0
 
@@ -83,14 +84,14 @@ showUD2GF opts env sentence = do
         putStrLn "#sum, an extractive summary (tree built from interpreted nodes)"
         let sts0 = devtree2abstrees besttree0
         let sts1 = map (expandMacro env) sts0
-        ifOpt opts "at" $ unlines $ map prAbsTree sts1 
+        ifOpt opts "at" $ unlines $ map prAbsTree sts1
         let scrs = map (checkAbsTreeResult env) sts1
         ifOpt opts "tc" $ unlines $ map prCheckResult scrs
         let sts = [t | Just t <- map resultTree scrs]
         if null sts then return () else
           ifOpt opts "lin" (unlines $ map (("SUMMARY LIN: " ++) . linearizeTree env (actLanguage env)) sts)
     else return ()
-  
+
   let allnodes = allNodesRTree besttree0
       orig = length allnodes
       interp = length (devStatus (root besttree0))
@@ -102,7 +103,7 @@ showUD2GF opts env sentence = do
        completeSentences = div interp orig, -- either 1 or 0
        typecorrectSentences = min 1 (length ts)  -- 1 if type-correct, 0 if not
        }
-  
+
   return (ts,stat)
 
 
@@ -145,7 +146,7 @@ data CheckResult = CheckResult {
   }
  deriving Show
 
-prCheckResult cr = unlines $ 
+prCheckResult cr = unlines $
   case resultUnknowns cr of
     [] -> []
     uks -> [unwords $ "unknown words:" : map showCId uks]
@@ -172,7 +173,7 @@ data DevNode = DevNode {
   devStatus     :: [UDId],         -- indices of words used in the best abstree in DevTrees --- redundant
   devWord       :: String,         -- the original word
   devAbsTrees   :: [AbsTreeInfo],  -- trees constructed at this node, with types and used words
-  devLemma      :: String, 
+  devLemma      :: String,
   devPOS        :: String,
   devFeats      :: [UDData],
   devLabel      :: String,
@@ -215,7 +216,7 @@ addBackups0 tr@(RTree dn trs) = case map collectBackup (tr:trs) of  -- backups f
         }
     )
     (map fst (tail btrs))
-  
+
  where
 
   -- add backups to tree, update usage with the nodes used in the backups (if no backups, do nothing)
@@ -254,7 +255,7 @@ theAbsTreeInfo dt = case devAbsTrees (root dt) of
 
 -- split trees showing just one GF tree in each DevTree
 splitDevTree :: DevTree -> [DevTree]
-splitDevTree tr@(RTree dn trs) =
+splitDevTree tr@(RTree dn trs) = -- traceShow (startCategory env)
   [RTree (dn{devAbsTrees = [t]}) (map (chase t) trs) | t <- devAbsTrees dn]
  where
   chase (ast,(cat,usage)) tr@(RTree d ts) = case elem (devIndex d) usage of
@@ -284,7 +285,7 @@ pruneDevTree  tr@(RTree dn dts) = RTree dn{devAbsTrees = pruneCatGroups (groupCa
   prune usages grp = case grp of
     t:ts | any (\u -> all (\x -> elem x u) (usage t)) usages -> prune usages ts
     t:ts -> t : prune (usage t : usages) ts
-    _ -> grp  
+    _ -> grp
   pruneCatGroups = concatMap (prune [])
 
 -- function application to a given set of arguments when building up DevTree
@@ -302,6 +303,7 @@ data ArgInfo = ArgInfo {
   argFeats  :: [UDData],      -- features of its head word
   argTree   :: AbsTree        -- the GF tree built at that node
   }
+  deriving Show
 
 combineTrees :: UDEnv -> DevTree -> DevTree
 combineTrees env =
@@ -312,22 +314,62 @@ combineTrees env =
 
   comb :: DevTree -> DevTree
   comb tr@(RTree dn dts) = case map comb dts of
-    ts -> keepTrying (RTree dn ts)
+    ts -> traceNoPrint (prDevNode 3 . root) "built" $ pruneDevTree $ rankDevTree $ keepTryingNew (RTree dn ts)
 
+  -- Apply all possible functions and iterate doing the same on the new trees until there's no new trees
+  keepTryingNew :: DevTree -> DevTree
+  keepTryingNew tr = tryEvenMore fs tr
+    where fs = newFuns (allFunsLocal tr) tr
+
+  -- Apply all possible functions to all trees, both old and new. This is needlessly slow.
   keepTrying :: DevTree -> DevTree
   keepTrying tr = case tryCombine (allFunsLocal tr)  tr of
     tr' | devAbsTrees (root tr') /= devAbsTrees (root tr) -> keepTrying tr'
-    _ -> traceNoPrint (prDevNode 3 . root) "built" $ pruneDevTree $ rankDevTree tr
+    _ -> tr
+
+  -- Apply all possible functions to the GF trees that were created in the previous iteration
+  tryEvenMore :: [FunInfo] -> DevTree -> DevTree
+  tryEvenMore [] tr = tr
+  tryEvenMore fis tr@RTree{root=nd} = tryEvenMore (newFuns (allFunsLocal onlyNewTree) nextTr) nextTr
+    where 
+      -- The head only contains new trees that were created in the previous iteration
+      onlyNewTree = tr {root = nd { devAbsTrees = map funInfoToAbsTreeInfo fis}}
+      -- Add the new trees to the old ones. There shouldn't really be any duplicates now, so there's 
+      -- a bit of redundant checking going on here.
+      nextTr =  tryCombine fis tr
+
+{-
+Ideas
+make tryEvenMore more like an unfold returning [FunInfo] (or [ArgTreeInfo])
+deduplicate new funInfos prior to comparing them to old list
+
+
+Handle one-argument functions differently
+or even better, don't collapse the tree to the product of all immediate children and the head
+but search for each function which combinatons applicable
+
+Make pruneTrees less aggressive
+
+Prune by linearizing and verifying that it matches
+
+-}
+
 
   tryCombine :: [FunInfo] -> DevTree -> DevTree
-  tryCombine finfos tr = case finfos of
-    f:fs -> tryCombine fs (addAbsTree f tr)
-    [] -> tr
+  tryCombine finfos tr =
+    -- trace ("combining:\n" ++ unlines (map ((" - "++) . prRTree show . funTree) finfos)
+    --       ++ "\nwith:\n " ++ prLinesRTree (prDevNode 4)  tr
+    --       ++ "\nresult:\n " ++ prLinesRTree (prDevNode 4)  res) 
+          res
+    where res =
+            foldl (flip addAbsTree) tr finfos
+    --case finfos of
+    --  f:fs -> tryCombine fs (addAbsTree f tr)
+    --  [] -> tr
 
   allFunsLocal :: DevTree -> [FunInfo]
   allFunsLocal tr@(RTree dn ts) =
     [FunInfo f labtyp abstree usage |
-        (f,labtyp) <- allFunsEnv env,
 
         -- for head and each immediate subtree, build the list of its already built abstrees, each with type and label
         -- argalts :: [[Arg]] -- one list for root and for each subtree
@@ -345,12 +387,25 @@ combineTrees env =
               sequence (filter (not . null) [filter (\x -> all (flip notElem us) (argUsage x)) xs  | xs <- argss]),
 
         let argseqs (arg:args) = [x:xs | x <- arg, xs <- argseqsAfter (argUsage x) args],
-        
-        (abstree,usage) <- tryFindArgs f labtyp (argseqs argalts)
+
+        let allF = allFunsEnv env,
+        let argss = argseqs argalts,
+        (f,labtyp) <-
+          -- trace (devWord dn ++ ": head: " ++ show (length $ devAbsTrees dn) ++ ", total: " ++ show (length argss)) $
+          -- trace ("tr:\n " ++ prLinesRTree (prDevNode 4)  tr 
+          -- -- ++ "\n\tallFuns: " ++ show (length allF)
+          -- ++ "\n\targss:\n" ++ unlines (("\t " ++) .show . map ((,,)<$> argNumber <*> argCatLab <*> prRTree show . argTree) <$> argss)
+          -- )
+           allF,
+        (abstree,usage) <- tryFindArgs f labtyp argss
       ]
 
   tryFindArgs :: CId -> LabelledType -> [[ArgInfo]] -> [(AbsTree,[UDId])]
   tryFindArgs f labtyp@(valcat,catlabs) argss =
+    --  trace ("f: "++ show f ++ " catlabs: " ++ show (length catlabs) ++ ", argss: " ++ show (length argss) ++ ", max args: " ++ show (maximum $ map length argss)
+    --        ++ "\n\tcatlabs: " ++ show catlabs
+    --        ++ "\n\targss:\n" ++ unlines (("\t " ++) .show . map ((,,)<$> argNumber <*> argCatLab <*> prRTree show . argTree) <$> argss)
+    -- )
     [(abstree,usage) |
         args <- argss,
         xis  <- (argTypeMatches catlabs args),
@@ -372,7 +427,7 @@ combineTrees env =
     _ -> [[]]
 
   addAbsTree :: FunInfo -> DevTree -> DevTree
-  addAbsTree finfo tr@(RTree dn ts) =
+  addAbsTree finfo tr@(RTree dn ts) = -- trace (devWord dn ++ ": " ++ show (length $ devAbsTrees dn)) $ -- trace (show dn) $
     RTree dn{
       devAbsTrees = let
                    acu = (funTree finfo,(fst (funTyp finfo),funUsage finfo))
@@ -380,10 +435,26 @@ combineTrees env =
                  in
                  if elem acu dts  -- the same tree with the same usage of subtrees is not added again
                     ---- || length dts > 123 ---- TODO parameterize "beam" size
-                   then dts
-                   else acu:dts,
+                   then -- trace ("same " ++ prRTree show (fst acu) ++ ", "  ++ show (snd acu))
+                     dts
+                   else -- trace ("new " ++ prRTree show (fst acu) ++ ", "  ++ show (snd acu)) $
+                     acu:dts,
       devStatus = maximumBy (\x y -> compare (length x) (length y)) [devStatus dn, funUsage finfo]
       } ts
+
+  funInfoToAbsTreeInfo :: FunInfo -> AbsTreeInfo
+  funInfoToAbsTreeInfo finfo = (funTree finfo,(fst (funTyp finfo),funUsage finfo)) 
+
+  newFuns :: [FunInfo] -> DevTree -> [FunInfo]
+  newFuns fis (RTree dn rts) = -- trace ("All funs:\n" ++ showFis fis ++ "\nFiltered funs:\n" ++ showFis result)
+   result
+    where
+      acu finfo = (funTree finfo,(fst (funTyp finfo),funUsage finfo))
+      dts = devAbsTrees dn
+      result = filter ((`notElem` dts) . funInfoToAbsTreeInfo) $ nubBy ((==) `on` funInfoToAbsTreeInfo) fis
+      -- TODO: nub won't be needed when the tree direct idea is implemented
+      showFis = unlines . map ((" - "++) . showAbsTreeInfo . funInfoToAbsTreeInfo)
+      showAbsTreeInfo acu = prRTree show (fst acu) ++ ", "  ++ show (snd acu) 
 
 analyseWords :: UDEnv -> DevTree -> DevTree
 analyseWords env = mapRTree lemma2fun
@@ -395,7 +466,7 @@ analyseWords env = mapRTree lemma2fun
     }
    where
     (isUnknown,justWords) = getWordTrees (devLemma dn) (cats (devPOS dn))
-    
+
   cats pos = maybe [] (map (either (Left. fst) Right)) $ M.lookup pos (catsForPOS env)
 
   -- find all functions that are possible parses of the word in any appropriate category
@@ -430,7 +501,7 @@ udtree2devtree :: UDTree -> DevTree
 udtree2devtree = markClosest . initialize
 
  where
- 
+
   initialize tr@(RTree un uts) =
     RTree (DevNode {
       devStatus = [],
@@ -446,7 +517,7 @@ udtree2devtree = markClosest . initialize
       devIsUnknown = True
      }) (map initialize uts)
 
-  markClosest tr@(RTree dn dts) = 
+  markClosest tr@(RTree dn dts) =
     RTree (dn {
        devClosest = hardClosest (devIndex dn)  -- top node not dominated
       }) (map (mark (devIndex dn)) dts)
@@ -463,4 +534,3 @@ udtree2devtree = markClosest . initialize
     if udPosition ui == 1
       then nextUDId ui -- first word linked to next one --- does not work for one-word sentence
       else previousUDId ui  -- other words to previous ones, also works for the last word
-